+++ src/main/java/bunch/util/BunchUtilities.java
@@ -136,18 +136,12 @@ public static Graph toInternalGraph(bunch.api.BunchMDG bunchMDG)
   ArrayList al = new ArrayList(bunchMDG.getMDGEdges());
   Hashtable nodes = new Hashtable();
 
-    /**
-     * Ignore reflexive edges
-     */
     for (Object o : al) {
         bunch.api.BunchMDGDependency bmd = (bunch.api.BunchMDGDependency) o;
 
         ParserNode currentNode = null;
         ParserNode targetNode = null;
 
-        /**
-         * Ignore reflexive edges
-         */
         if (bmd.getSrcNode().equals(bmd.getDestNode()))
             continue;
 
@@ -177,7 +171,7 @@ public static Graph toInternalGraph(bunch.api.BunchMDG bunchMDG)
             //System.out.println("Adding weight " + w);
         } else {
             Integer wExisting = (Integer) currentNode.dWeights.get(dep);
-            Integer wtemp = w.intValue() + wExisting.intValue();
+            Integer wtemp = w + wExisting;
             currentNode.dWeights.put(dep, wtemp);
         }
 
@@ -186,7 +180,7 @@ public static Graph toInternalGraph(bunch.api.BunchMDG bunchMDG)
             targetNode.beWeights.put(src, w);
         } else {
             Integer wExisting = (Integer) targetNode.beWeights.get(src);
-            Integer wtemp = w.intValue() + wExisting.intValue();
+            Integer wtemp = w + wExisting;
             targetNode.beWeights.put(src, wtemp);
         }
 
@@ -219,7 +213,7 @@ public static Graph toInternalGraph(bunch.api.BunchMDG bunchMDG)
     ParserNode p = (ParserNode)nl[i];
     n.setName(p.name);
     Integer nid = (Integer)nameTable.get(p.name);
-    n.nodeID = nid.intValue();
+    n.nodeID = nid;
     n.dependencies = ht2ArrayFromKey(nameTable,p.dependencies);
     n.weights = ht2ArrayValFromKey(p.dWeights);
     n.backEdges = ht2ArrayFromKey(nameTable,p.backEdges);
@@ -241,7 +235,7 @@ private static int[] ht2ArrayFromKey(Hashtable key, Hashtable values) {
       for(int i = 0; i < oa.length; i++) {
         String s = (String)oa[i];
         Integer val = (Integer)key.get(s);
-        retArray[i] = val.intValue();
+        retArray[i] = val;
       }
       return retArray;
     } catch(Exception e) {
@@ -262,7 +256,7 @@ private static int[] ht2ArrayValFromKey(Hashtable values) {
       {
         String s = (String)oa[i];
         Integer value = (Integer)values.get(s);
-        retArray[i] = value.intValue();
+        retArray[i] = value;
       }
       return retArray;
     } catch(Exception e) {
